+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased][unreleased]
 ### Added
 - Implement Parquet format. (Pau Tallada)
+- Patch comments and column names in Parquet format. (Pau Tallada)
 
 
 ## [2.6.1] - 2021-11-17